requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Marshmallow==1.2.1
psycopg2-binary==2.9.10
lxml==5.3.0
python-dotenv==1.0.1
marshmallow-sqlalchemy==1.1.0
pytest==8.3.4
pytest-flask==1.3.0